Add parser module
